Refatorando showTitleScreen()...
