+++ titleScreen.py
@@ -1,0 +1,63 @@
+from time import sleep
+from os import system
+
+def showTitleScreen():
+    frames = []
+    frames.append('''
+  /$$$$$$   /$$$$$$   /$$$$$$   /$$$$$$
+ /$$__  $$ /$$$_  $$ /$$__  $$ /$$$_  $$
+|__/  \ $$| $$$$\ $$|__/  \ $$| $$$$\ $$
+  /$$$$$$/| $$ $$ $$  /$$$$$$/| $$ $$ $$
+ /$$____/ | $$\ $$$$ /$$____/ | $$\ $$$$
+| $$      | $$ \ $$$| $$      | $$ \ $$$
+| $$$$$$$$|  $$$$$$/| $$$$$$$$|  $$$$$$/
+|________/ \______/ |________/ \______/
+''')
+    frames.append('''
+ $$$$$$\   $$$$$$\   $$$$$$\   $$$$$$\\
+$$  __$$\ $$$ __$$\ $$  __$$\ $$$ __$$\\
+\__/  $$ |$$$$\ $$ |\__/  $$ |$$$$\ $$ |
+ $$$$$$  |$$\$$\$$ | $$$$$$  |$$\$$\$$ |
+$$  ____/ $$ \$$$$ |$$  ____/ $$ \$$$$ |
+$$ |      $$ |\$$$ |$$ |      $$ |\$$$ |
+$$$$$$$$\ \$$$$$$  /$$$$$$$$\ \$$$$$$  /
+\________| \______/ \________| \______/
+''')
+
+    frames.append('''
+  ______    ______    ______    ______
+ /      \  /      \  /      \  /      \\
+|  $$$$$$\|  $$$$$$\|  $$$$$$\|  $$$$$$\\
+ \$$__| $$| $$$\| $$ \$$__| $$| $$$\| $$
+ /      $$| $$$$\ $$ /      $$| $$$$\ $$
+|  $$$$$$ | $$\$$\$$|  $$$$$$ | $$\$$\$$
+| $$_____ | $$_\$$$$| $$_____ | $$_\$$$$
+| $$     \ \$$  \$$$| $$     \ \$$  \$$$
+ \$$$$$$$$  \$$$$$$  \$$$$$$$$  \$$$$$$
+''')
+
+    frames.append('''
+  ______    ______    ______    ______
+ /      \  /      \  /      \  /      \\
+/$$$$$$  |/$$$$$$  |/$$$$$$  |/$$$$$$  |
+$$____$$ |$$$  \$$ |$$____$$ |$$$  \$$ |
+ /    $$/ $$$$  $$ | /    $$/ $$$$  $$ |
+/$$$$$$/  $$ $$ $$ |/$$$$$$/  $$ $$ $$ |
+$$ |_____ $$ \$$$$ |$$ |_____ $$ \$$$$ |
+$$       |$$   $$$/ $$       |$$   $$$/
+$$$$$$$$/  $$$$$$/  $$$$$$$$/  $$$$$$/
+''')
+
+    timer = 0.0   # Seconds
+    animationDuration = 4.0   # Seconds
+    delayBetweenFrames = 0.2   # Seconds
+    frameCounter = 0   # Frames
+    frameAmount = len(frames)   # Frames
+
+    while(timer < animationDuration):
+        currentFrameIndex = frameCounter % frameAmount
+        print(frames[currentFrameIndex])
+        frameCounter += 1
+        sleep(delayBetweenFrames)
+        system("cls")
+        timer += delayBetweenFrames
